+++ modules/apps/custom-elements/custom-elements-api/bnd.bnd
@@ -2,7 +2,6 @@ Bundle-Name: Liferay Custom Elements API
 Bundle-SymbolicName: com.liferay.custom.elements.api
 Bundle-Version: 1.0.0
 Export-Package:\
-	com.liferay.custom.elements.constants,\
 	com.liferay.custom.elements.exception,\
 	com.liferay.custom.elements.model,\
 	com.liferay.custom.elements.service,\
