--- modules/apps/custom-elements/custom-elements-api/src/main/resources/com/liferay/custom/elements/constants/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
