+++ modules/apps/custom-elements/custom-elements-web/src/main/java/com/liferay/custom/elements/web/internal/portlet/action/CustomElementsViewMVCRenderCommand.java
@@ -14,7 +14,7 @@
 
 package com.liferay.custom.elements.web.internal.portlet.action;
 
-import com.liferay.custom.elements.constants.CustomElementsPortletKeys;
+import com.liferay.custom.elements.web.internal.constants.CustomElementsPortletKeys;
 import com.liferay.custom.elements.web.internal.constants.CustomElementsWebKeys;
 import com.liferay.custom.elements.web.internal.display.context.CustomElementsDisplayContext;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
@@ -31,7 +31,7 @@ import org.osgi.service.component.annotations.Component;
 @Component(
 	immediate = true,
 	property = {
-		"javax.portlet.name=" + CustomElementsPortletKeys.CUSTOM_ELEMENTS_ADMIN,
+		"javax.portlet.name=" + CustomElementsPortletKeys.CUSTOM_ELEMENTS,
 		"mvc.command.name=/"
 	},
 	service = MVCRenderCommand.class
