+++ modules/apps/custom-elements/custom-elements-web/src/main/java/com/liferay/custom/elements/web/internal/portlet/action/DeleteCustomElementsSourceMVCActionCommand.java
@@ -14,8 +14,8 @@
 
 package com.liferay.custom.elements.web.internal.portlet.action;
 
-import com.liferay.custom.elements.constants.CustomElementsPortletKeys;
 import com.liferay.custom.elements.service.CustomElementsSourceLocalService;
+import com.liferay.custom.elements.web.internal.constants.CustomElementsPortletKeys;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
 import com.liferay.portal.kernel.util.ParamUtil;
@@ -33,7 +33,7 @@ import org.osgi.service.component.annotations.Reference;
 @Component(
 	immediate = true,
 	property = {
-		"javax.portlet.name=" + CustomElementsPortletKeys.CUSTOM_ELEMENTS_ADMIN,
+		"javax.portlet.name=" + CustomElementsPortletKeys.CUSTOM_ELEMENTS,
 		"mvc.command.name=/custom_elements/delete_custom_elements_source"
 	},
 	service = MVCActionCommand.class
