+++ modules/apps/custom-elements/custom-elements-web/src/main/java/com/liferay/custom/elements/web/internal/portlet/action/EditCustomElementsSourceMVCActionCommand.java
@@ -14,9 +14,9 @@
 
 package com.liferay.custom.elements.web.internal.portlet.action;
 
-import com.liferay.custom.elements.constants.CustomElementsPortletKeys;
 import com.liferay.custom.elements.model.CustomElementsSource;
 import com.liferay.custom.elements.service.CustomElementsSourceLocalService;
+import com.liferay.custom.elements.web.internal.constants.CustomElementsPortletKeys;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
 import com.liferay.portal.kernel.service.ServiceContext;
@@ -37,7 +37,7 @@ import org.osgi.service.component.annotations.Reference;
 @Component(
 	immediate = true,
 	property = {
-		"javax.portlet.name=" + CustomElementsPortletKeys.CUSTOM_ELEMENTS_ADMIN,
+		"javax.portlet.name=" + CustomElementsPortletKeys.CUSTOM_ELEMENTS,
 		"mvc.command.name=/custom_elements/edit_custom_elements_source"
 	},
 	service = MVCActionCommand.class
