+++ modules/apps/custom-elements/custom-elements-web/src/main/resources/content/Language.properties
@@ -1,3 +1,3 @@
 add-custom-elements-source=Add Custom Element
 html-element-name=HTML Element Name
-javax.portlet.title.com_liferay_custom_elements_web_portlet_CustomElementsPortlet=Custom Elements
+javax.portlet.title.com_liferay_custom_elements_web_internal_portlet_CustomElementsPortlet=Custom Elements
